==> notebooks/3.description_kmeans_clustering.ipynb <==
# Notebook 3 - KMeans Clustering & Exploratory Analysis
Once vectorization is complete, a KMeans clustering model is used to group similar bodies of text together.  Once grouped, we totaled the number of trades per cluster per INN (Russian Tax ID number).   Next, we totaled the number of trades per cluster for all known Russian Arms Exporters, and converted those totals to ratios, where each ratio represents the percentage of trades that fall into that cluser.

We are building a 'profile' for known Russian arms exporters.  Based on analysis of the text content of individual trades, known Russian arms exporters can be assigned to  clusters.  Our final product will use this known arms exporter profile to compare against new trade data.  It aims to answer the questions: how many companies trade similar products/in a similar manner as known arms exporters, and how similar are they?

It assigns a 'similarity score' for each INN in the dataset using a measure of inverse euclidian distance.  In simpler terms, it checks how 'similar' each group of percentages is to the percentages of all known Russian arms exporters.

# required pip installations
!pip install --upgrade pip
!pip install joblib
!pip install --upgrade s3fs
!pip install googletrans

# for dask machine learning cluster
!pip install dask-ml
!pip install executor

# IMPORTS

# dataframe
import dask.dataframe as dd
import pandas as pd

# machine learning/analysis
#from sklearn.cluster import KMeans
import dask_ml.cluster as dask_ml_model # sklearn's skmeans took up too much memory to run.
from sklearn.model_selection import train_test_split

# saving model to S3 bucket
import tempfile
import boto3
import joblib

# translate results
from googletrans import Translator
#import executor as e

# read df_trade_desc_filt_vector for nlp
df = dd.read_csv('s3://labs20-arms-bucket/data/df_train_description_filtered_vectorizedIF2.csv',dtype={'CONSIGNOR_INN': 'object'})
#                 dtype={'CONSIGNOR_INN': 'str', 'Unnamed: 0': 'str'}, usecols=range(0, 312))

df = df.drop(columns=['Unnamed: 0'])
# remove all ',' from CONSIGNOR_INN column
# somehow was missed in REGEX filter from cleaning_trade_data notebook
#df['CONSIGNOR_INN'] = df['CONSIGNOR_INN'].str.replace(',', '')

df.head()

# read dataframe to memory
# allows pandas operations to be performed
df = df.compute()

type(df)

### Feeding Data Into the Model
Similar to the TfidfVectorizer, our dask_ml_model.KMeans model object must be trained on an array.  In our inital analysis we did not perform hyperparameter tuning, and selected 10 clusters for our KMeans cluster as a default.  Once the array was created, we fit the model on the array, and extracted model.labels_ to use as our cluster names, and added the results to our dataframe.

# variable manipulation to feed into KMeans model
# pull create variable containing dataframe of vectorized words only, all rows, columns indexed 4 and onward
X = df.iloc[:, 3:]

# check dataframe to confirm its columns only contain word vectors
X.head()

# check dataframe to confirm its columns only contain word vectors
X.shape

# convert X dataframe into array
# necessary to feed to KMeans model
X_array = X.values

### KMeans Model Selection
Because sklearn's KMeans module used too much memory, Labs20 group decided to use dask_ml_model from dask.clusters library.  This is because the default initializer for dask_ml_model.KMeans is `k-means||`, compared to `k-means++` from scikit-learn.  `k-means||` is designed to work well in a distributed environment such as SageMaker, whereas `k-means++` reads everything to memeory at once and can cause errors.

# define KMeans model
# n_jobs = -1 tells model to use all available processors
model = dask_ml_model.KMeans(n_clusters=10)

# fit model on vectorized word array
model.fit(X_array)

# Once the model was trained on our array of word vectors, we pickled it to our S3 bucket for use in our final product.
s3 = boto3.resource('s3')
bucket=s3.Bucket('labs20-arms-bucket')
key = "modelf.pkl"

# WRITE/SAVE 'model' to s3 bucket
with tempfile.TemporaryFile() as fp:
    joblib.dump(model, fp, compress=3)
    fp.seek(0)
    bucket.put_object(Body=fp.read(), Key=key)

# test READ/LOAD of model from S3 bucket
with tempfile.TemporaryFile() as fp:
    bucket.download_fileobj(Fileobj=fp, Key=key)
    fp.seek(0)
    model_load = joblib.load(fp)

# Confirming model saved to S3 bucket is the same as model created in notebook
model_load

# Confirming vmodel saved to S3 bucket is the same as model created in notebook
model

# Define model labels in variable 'labels'
labels = model.labels_

#Glue back to originaal data
df['cluster'] = labels

# reduce dataframe to necessary columns, no longer need text now that we have clusters
df = df[['CONSIGNOR_NAME', 'CONSIGNOR_INN', 'PROCESSED_TEXT', 'cluster']]
df.head()

# Export dataframe with cluster assignments to S3 Bucket
df.to_csv('s3://labs20-arms-bucket/data/df_train_with_clustersIF2.csv')

### Initial Analysis
Once our model was saved, we recreated the 'tokenize' function and loaded the pickled vectorizer to view the results of our clustering. For some reason we could not load the pickled 'tokenize' function from our S3 bucket, so we had to recreate it each time.  Because the tokenize function was used in our vectorizer, it must be loaded into the memory of the notebook before the vectorizer is loaded from the S3 bucket, otherwise an error message will generate.

To see how our model/vectorizer performed we created a for loop that generated the top 10 terms per cluster and translated them to English.  To our delight, the clusters seemed to organize the text corpuses around specific industries; clusters 0, 6, and 7 even displayed words strongly associated with arms exports and heavy machine building!  This is a sign that our model grouped similar trades correctly.

# create S3 object
s3 = boto3.resource('s3')
bucket=s3.Bucket('labs20-arms-bucket')

# load vectorizer for analysis
key = "vectorizerf.pkl"
# test READ/LOAD of vectorizer from S3 bucket
with tempfile.TemporaryFile() as fp:
    bucket.download_fileobj(Fileobj=fp, Key=key)
    fp.seek(0)
    vectorizer = joblib.load(fp)

# should not run the code many time because the googleapi will give you 'JSONDecodeError: Expecting value: line 1 column 1 (char 0)'
# This for loop generates the top 10 terms per cluster and translates them to English
translator = Translator()
print("Top terms per cluster:")
order_centroids = model.cluster_centers_.argsort()[:, ::-1]
terms = vectorizer.get_feature_names()
for i in range(model.n_clusters):
    print("Cluster %d:" % i),
    for ind in order_centroids[i, :15]:
        print("Russian:",' %s' % terms[ind], ", English:", translator.translate('{}'.format(terms[ind])).text)
    print

# Create listAllcluster list to count the number of trade text corpuses falling into each cluster for all INN numbers.
listAllcluster = df.cluster.value_counts()
listAllcluster

### Further Exploration/Known Russian Arms Exporters
We now introduce known arms exporter INNs into our analysis.  Unfortunately, the list containing these INNs was manually put together by Labs16 & Labs20 groups, and there is neither a database containing this information or a way to automatically scrape them into our notebook.  The GoogleSpreadsheet containing the running list of known Russian arms exporter INNs can be found here: https://docs.google.com/spreadsheets/d/1-RDS-STLXPQ3tMkPe4hXwt7CL_fM8IrgRWn0uDX6IF4/edit?usp=sharing

As stated in Notebook 3, the goal here is to create a 'profile' for known arms exporters.  This is done by totalling the trades per cluster for all INNs.  Once totalled, the known exporter totals dataframe will be exported to our S3 bucket and used as a comparision for our product's final analysis.

# This cell saves the string list of known arms exporter inns as variable inn_arms_exp_total
inn_arms_exp_total = ['7718852163',  '7740000090',    '7731084175',  '6161021690',
                      '3807002509',  '6672315362',    '7802375335',  '7813132895',  
                      '7731280660',  '7303026762',    '5040007594',  '2501002394',  
                      '7807343496',  '7731559044',    '5042126251',  '7731595540',    
                      '7733018650',  '7722016820',    '7705654132',  '7714336520',    
                      '7801074335',  '6229031754',    '7830002462',  '6825000757',  
                      '5043000212',  '7802375889',    '5010031470',  '1660249187',  
                      '7720015691',  '6154573235',    '5038087144',  '7713006304',  
                      '7805326230',  '5023002050',    '4007017378',  '7714013456',  
                      '17718852163', '7811406004',    '7702077840',  '7839395419',  
                      '7702244226',  '7704721192',    '7731644035',  '7712040285',
                      '7811144648',  '4345047310',    '7720066255',  '6607000556',
                      '1832090230',  '1835011597',    '3305004083',  '4340000830',
                      '5074051432',  '1841015504',    '7105008338',  '7106002829', 
                      '7704274402',  '5942400228',    '7105514574',  '5012039795', 
                      '7714733528',  '3904065550',    '6825000757',  '7807343496', 
                      '7731559044',  '7805231691',    '7704859803',  '0273008320',
                      '7704274402',  '2902059091',    '7805034277',  '7727692011',
                      '7733759899',  '6154028021',    '7328032711',  '2635002815',
                      '5040097816',  '5027033274',    '5250018433',  '5200000046',
                      '7743813961',  '7718016666',    '5047118550',  '7704274402']

We created a simple `for` loop to check how many times our known INN numbers appeared in the dataset.

# create list of predicted inns of all trades assigned to all clusters
predicted_INNs_total = list(df['CONSIGNOR_INN'])

# create 'for' loop to see how manny times know INNs show up in predicted_INNs list
# some INNS are very present in clusters 4 and 7, others not so much
# expanding list of known arms exporters would be very helpful for this portion
for i in inn_arms_exp_total:
    print("INN:", i, "  " ,"number of occurances in training dataset (all clusters):", predicted_INNs_total.count(i))

### Final Analysis
We then totaled the trade-count-per-cluster for ALL known Russian Arms exporters, transposed it, and saved it to a dataframe.  This transposed dataset will be used as the final comparision in our product.  For example:

| Cluster% | Known Arms Exporters% | INN in Dataset% |
| --- | --- | --- |
| % of trades falling into cluster0 | 30% | 26% |
| % of trades falling into cluster1 | 12% | 6% |
| % of trades falling into cluster2 | 5% | 15% |
| % of trades falling into cluster3 | 5% | 0% |
| % of trades falling into cluster4 | 5% | 2% |
| % of trades falling into cluster5 | 20% | 18% |
| % of trades falling into cluster6 | 5% | 3% |
| % of trades falling into cluster7 | 1% | 12% |
| % of trades falling into cluster8 | 3% | 0% |
| % of trades falling into cluster9 | 5% | 15% |

# Create subset of dataframe containing trade corpuses of known arms exporters only
X = df[df['CONSIGNOR_INN'].isin(inn_arms_exp_total)]

# Convert known arms exporter data
listArmcluster = pd.DataFrame(X.cluster.value_counts())
# add 0 value for cluster 9 index, as no known arms exporter trades fell into cluster 9
listArmcluster.loc[9]=0
listArmcluster

# arms cluster fixing
listArmcluster = listArmcluster.sort_index().reset_index()
listArmcluster = listArmcluster.rename(columns={'index': "cluster", 'cluster': "known_AE_trade_count"})
listArmcluster = listArmcluster.T.reset_index().rename(columns={'index':'CONSIGNOR_INN', 0:'clust0', 1:'clust1', 2:'clust2',
                                                    3:'clust3', 4:'clust4', 5:'clust5', 6:'clust6',7:'clust7', 8:'clust8', 9:'clust9'}).rename_axis('', axis=1)
listArmcluster = listArmcluster.drop([0])

# 'profile' for russian arms exporters
listArmcluster

# export listAllcluster to S3 bucket
listArmcluster.to_csv('s3://labs20-arms-bucket/data/armsclustersf.csv')
