==> notebooks/7.results.ipynb <==
# Notebook 7 - Result Exploration

import pandas as pd

# Load results.csv, saved in function notebook
df = pd.read_csv('s3://labs20-arms-bucket/data/results.csv').drop(columns=['Unnamed: 0'])

df.sort_values(by='pdist_score', ascending=False).head(100)
